requirements.txt
django>=5.1.7
azure-data-tables>=12.6.0
azure-storage-blob>=12.25.1
